requirements.txt
Markdown>3.0.0
PyYAML>5.0
rdflib>4.2
rdflib-jsonld>0.4.0
